[MINOR] Add basic tests and setup
