+++ .eslintrc.json
@@ -2,7 +2,8 @@
   "env": {
     "node": true,
     "commonjs": true,
-    "es2021": true
+    "es2021": true,
+    "jest": true
   },
   "extends": "eslint:recommended",
   "parserOptions": {
