+++ .github/workflows/node-ci.yml
@@ -28,3 +28,6 @@ jobs:
 
       - name: Lint Source Files
         run: npm run lint
+
+      - name: Run Tests
+        run: npm test
